add decision point groups package, rearrange modules
