+++ src/ssvc/dp_groups/cvss/__init__.py
@@ -1,0 +1,14 @@
+#!/usr/bin/env python
+"""
+file: __init__.py
+author: adh
+created_at: 9/20/23 4:54 PM
+"""
+
+
+def main():
+    pass
+
+
+if __name__ == "__main__":
+    main()
